Fix vote type from int to double

+++ core/src/main/java/com/ataulm/mystories/discovertv/GsonDiscoverTv.java
@@ -36,7 +36,7 @@ class GsonDiscoverTv {
         double popularity;
 
         @SerializedName("vote_average")
-        int voteAverage;
+        double voteAverage;
 
         @SerializedName("vote_count")
         int voteCount;
